Totally remove jvm version
What the "official" way to install or the current aur
official package do is essentially the same:
  use coursier to download the binary built and existing in gh.

the coursier way needs coursier and creates an extra useless bash script
and changes the jar to have "release" info for coursier itself to update
it. Which nobody wants
in an system package.

So the official way is already using the pre compiled binary and the jvm
version is dead. Instead of using coursier to do it, the native version
I have just uses it directly form github. It's for the best.